TypeSctipt for ConnectionStatus, ConsoleContent, DBCAttribute, DBCComment
