+++ src/control/ConnectionStatus/index.ts
@@ -1,0 +1,37 @@
+import { BaseControl, NQDOM } from "webnetq-js";
+// @ts-ignore
+import { SIGNAL_STATE_TEXT, SIGNAL_STATE_ON, SIGNAL_STATE_OFF } from "uictmplt-loader!./template.ts";
+
+export class ConnectionStatus extends BaseControl {
+  private _valueElm?: HTMLElement;
+  private _textElm?: HTMLElement;
+
+  protected _init() {
+    this._valueElm = NQDOM.getElementByClassName(this.element, SIGNAL_STATE_OFF);
+    this._textElm = NQDOM.getElementByClassName(this.element, SIGNAL_STATE_TEXT);
+  }
+
+  public get value(): boolean {
+    return !!this._valueElm && this._valueElm.classList.contains(SIGNAL_STATE_ON);
+  }
+
+  public set value(value: boolean) {
+    if (this._valueElm) {
+      this._valueElm.classList.toggle(SIGNAL_STATE_OFF, !value);
+      this._valueElm.classList.toggle(SIGNAL_STATE_ON, value);
+    }
+  }
+
+  public get text(): string {
+    return this._textElm?.textContent || "";
+  }
+
+  public set text(value: string) {
+    this._textElm && (this._textElm.textContent = value);
+  }
+
+  public setState(value: boolean, text: string) {
+    this.value = value;
+    this.text = text;
+  }
+};
