+++ src/control/ConsoleContent/index.ts
@@ -1,7 +1,4 @@
 import { BaseControl } from "webnetq-js";
-import { } from 'uictmplt-loader!./template.mjs';
 
 export class ConsoleContent extends BaseControl {
-  _init() {
-  }
 };
