+++ src/control/ConsoleContent/template.ts
@@ -1,8 +1,8 @@
-import ControlMaker from '../../lib/ControlMaker.mjs';
+import ControlMaker from "@/lib/ControlMaker.mjs";
 
-const mk = new ControlMaker('ConsoleContent', import.meta.url);
+const mk = new ControlMaker("ConsoleContent");
 
-const clss = mk.newClassNameMap([
+const clss: any = mk.newClassNameMap([
   "ROOT_CLASS",
 ]);
